-added some commented out routing code to help on how page navigation works \n-moved dropdown code to its own folder
