+++ src/app/app-routing.module.ts
@@ -1,10 +1,17 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { RouterModule, Routes } from '@angular/router';
+
+import { RegisterComponent } from './register/register.component';
+import { LoginComponent } from './login/login.component';
+
+const routes: Routes = [
+  // { path: '', redirectTo: '/#', pathMatch: 'full' },
+  { path: 'sign-up', component: RegisterComponent },
+  { path: 'login', component: LoginComponent },
+];
 
 @NgModule({
-  imports: [
-    CommonModule
-  ],
-  declarations: []
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule]
 })
 export class AppRoutingModule { }
