+++ src/app/app.component.html
@@ -1,12 +1,5 @@
-<!--The content below is only a placeholder and can be replaced.-->
-<body>
-    <app-dropdown></app-dropdown>
+<!-- <router-outlet></router-outlet> -->
+
+  <app-dropdown></app-dropdown>
   <img class="logo" src="../assets/rsi-logo.jpg" alt="RSI Logo" />
   <app-image-result></app-image-result>
-  </body>
-
-
-
-
-
-
