+++ src/app/app.component.ts
@@ -4,11 +4,10 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
- styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.css']
 })
 export class AppComponent {
   title = 'RSI ABQ Meme Generator';
-
 }
